+++ CHANGELOG.md
@@ -9,3 +9,10 @@
 ## 1.0.2
 
 - Added new models to `ChatModel` enum in `enum_ai_models.dart`.
+
+## 2.0.0
+
+- Added new vision models to `ChatVisionModel` enum in `enum_ai_models.dart`. Introduced simple conversation memory.
+- Major change to error handling. No longer returns null.
+- Removed deprecated models. See `https://docs.together.ai/docs/deprecations` for now info.
+- Please make sure to review the model names from the website.